Mapping out endpoints via inspector
Splitting UnrestEndpoing>>endpointPath b/w instance and class sides.
We use the latter for determining the base path of an endpoint. The instance
side method simply refers to the class side method, but allows specialization
via inheritance. This way, when visualizing endpoints - we get a path w/o needing
to instantiate the endpoint class, and for classes that spcialize the path on the 
instance side - we really just need the base path, when it comes to visualizing endpoints
and their paths