FIx issue with genereating cURL
When generating cURL command, we're not actually executing the endpoint's executing method, and thus won't have access to its HTTP method and path, which are determined from the method's pragmas. To remedy this, the method that's calling #curlForExecuting:with: will pass a block that makes up for the absence of contextual information. This needs to be executed before calling #prepareForExecutingOn: