Fix generation of curl statements
We need to actually call the executing method of an endpoint to have a properly configured ZnClient. But we don't want to actually execute the request. To make that happen, we stub out some methods in the client to return curl string instead of executing a request when calling ZnClient>>#execute