Simplify path handling
Classes are no longer required to define a path. The class side #endpointPath was removed, 
but the instance side stayed. Final path is determined by first resolving the endpointPath 
against WSClient's baseUrl - if not nil; then resolving the executing method's endpoint 
accessor's path - captured in the pragma of that method.

- Use single pragmas for specifying path (i.e. <get> or <get: '/path/somewhere'>
- The old <path> is still valid and used when a no-argument <METHOD> pragma exists
- Adding WSEndpointAccessor to capture the endpoint pragma values
- Moving path resolution to WSClient so as not to burden TWSEndpoint
- Renamed #supportedHttpMethods to #httpMethods