Switch from instance endpointPath to class pathPrefix
We'd like to know the final path from a method without instantiating the class.
Also, having this on the instance side invites an idea of including object state in composition of the path,
which would be problematic if we wanted to know paths w/o instantiating the client or the endpoint.