Use path encoders for encoding endpoint paths
Not all API implementations distringuish logic endpoints using a path. It is not uncommon to find an API that takes a unique identifier in a query parameter. But this could be done in a number of other ways. We still would like to distringuish between logic endpoints, but when it comes to using the pluggable endpoint - which is typically derived using a path - we need to figure out how to encode the unique 'path', or what have you...

Using #ignoresPath approach is clunky. Instead, use an encoder object. The behavior is captured in TWSPathEncoder. WSPluggableEndpoint uses that trait and implements the default behavior of resolving the path as a URL path.