Pull out #baseUrl from WSClient's requirements
WSClient needs to respond to #baseUrl, but it has no obligation to  keep that in an ivar. In fact, this is best to be left to the implementation. For example - maybe we'd like to store a larger object like Domain or Configuration of some sort, from which #baseUrl can be determined.